# docs/API.yaml
openapi: 3.0.3
info:
  title: Tasky API
  description: |
    REST API for Tasky — Mongolia's most trusted domestic service marketplace.

    ## Authentication
    All endpoints (unless noted) require a Bearer JWT in the `Authorization` header.
    Obtain tokens via the OTP login flow (`/auth/otp/*`).

    ## Pagination
    All list endpoints use **cursor-based pagination**.
    Pass `cursor` (opaque string) and `limit` (default 20, max 100) as query parameters.
    Responses wrap results in `{ data: [], cursor: { next, has_more } }`.

    ## Errors
    All errors return a standard JSON body: `{ code, message, trace_id }`.

    ## Currency
    All monetary amounts are integers in MNT (Mongolian Tugrik).

    ## Idempotency
    Critical state-changing endpoints require an `Idempotency-Key` header.
    Payment and wallet endpoints are documented as post-MVP deferred surfaces.
  version: 1.1.0
  contact:
    name: Tasky Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Opaque cursor for pagination. Pass the `next` value from a previous response.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (default 20, max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent handling of critical state-changing requests.
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          example: TASK_ALREADY_BOOKED
        message:
          type: string
          example: This task has already been assigned to another tasker.
        trace_id:
          type: string
          example: abc-123-def
    CursorPagination:
      type: object
      required: [next, has_more]
      properties:
        next:
          type: string
          nullable: true
          description: Opaque cursor to fetch the next page. Null if no more results.
          example: eyJpZCI6MTAwfQ==
        has_more:
          type: boolean
          example: true
...
    Profile:
      type: object
      required: [id, phone, role, status, full_name, avatar_url, rating_avg, completed_tasks, is_pro, created_at]
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "+97699001122"
        role:
          type: string
          enum: [CUSTOMER, TASKER, ADMIN]
        status:
          type: string
          enum: [PENDING, VERIFIED, BANNED, SUSPENDED]
        full_name:
          type: string
          example: Бат-Эрдэнэ
        avatar_url:
          type: string
          format: uri
          nullable: true
        rating_avg:
          type: number
          format: double
          example: 4.7
        completed_tasks:
          type: integer
          example: 12
        is_pro:
          type: boolean
          description: Auto-assigned when completed_tasks > 5 AND rating_avg > 4.5.
        created_at:
          type: string
          format: date-time
    StrikePolicy:
      type: object
      required:
        - strikeWindowDays
        - strikeThreshold
        - firstSuspensionDays
        - repeatSuspensionDays
        - repeatOffenseWindowDays
        - autoUnsuspendEnabled
        - updatedAt
      properties:
        strikeWindowDays:
          type: integer
          minimum: 1
          maximum: 365
        strikeThreshold:
          type: integer
          minimum: 1
          maximum: 10
        firstSuspensionDays:
          type: integer
          minimum: 1
          maximum: 365
        repeatSuspensionDays:
          type: integer
          minimum: 1
          maximum: 365
        repeatOffenseWindowDays:
          type: integer
          minimum: 1
          maximum: 730
        autoUnsuspendEnabled:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    StrikePolicyUpdateRequest:
      type: object
      required:
        - strikeWindowDays
        - strikeThreshold
        - firstSuspensionDays
        - repeatSuspensionDays
        - repeatOffenseWindowDays
        - autoUnsuspendEnabled
      properties:
        strikeWindowDays:
          type: integer
          minimum: 1
          maximum: 365
        strikeThreshold:
          type: integer
          minimum: 1
          maximum: 10
        firstSuspensionDays:
          type: integer
          minimum: 1
          maximum: 365
        repeatSuspensionDays:
          type: integer
          minimum: 1
          maximum: 365
        repeatOffenseWindowDays:
          type: integer
          minimum: 1
          maximum: 730
        autoUnsuspendEnabled:
          type: boolean

    # ─── Verification ─────────────────────────────────────────
...
    Category:
      type: object
      required: [id, name, name_mn, icon_url, is_active, sort_order]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Cleaning
        name_mn:
          type: string
          example: Цэвэрлэгээ
        icon_url:
          type: string
          format: uri
        is_active:
          type: boolean
        sort_order:
          type: integer

    # ─── Tasks ────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, category_id, description, budget, location_lat, location_lng, location_text, status, scheduled_at, photos, created_at]
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category:
          $ref: '#/components/schemas/Category'
        customer_id:
          type: string
          format: uuid
        description:
          type: string
        budget:
          type: integer
          description: Fixed budget in MNT.
          example: 50000
        location_lat:
          type: number
          format: double
        location_lng:
          type: number
          format: double
        location_text:
          type: string
          description: Exact address (visible only to task owner and booked tasker).
          example: Энхтайваны өргөн чөлөө 5-р байр, 301 тоот
        status:
          type: string
          enum: [OPEN, ASSIGNED, COMPLETED, CANCELLED]
        scheduled_at:
          type: string
          format: date-time
        photos:
          type: array
          items:
            type: object
            properties:
              storage_key:
                type: string
              url:
                type: string
                format: uri
              sort_order:
                type: integer
          maxItems: 3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Booking:
      type: object
      required: [id, task_id, tasker_id, customer_id, price, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        task:
          $ref: '#/components/schemas/Task'
        tasker_id:
          type: string
          format: uuid
        tasker:
          $ref: '#/components/schemas/Profile'
        customer_id:
          type: string
          format: uuid
        customer:
          $ref: '#/components/schemas/Profile'
        price:
          type: integer
          description: Agreed price in MNT.
        status:
          type: string
          enum: [ASSIGNED, COMPLETED, CANCELLED]
        cancellation_fee:
          type: integer
          nullable: true
          description: Post-MVP monetization field. Null in phase-1.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ─── Wallet ───────────────────────────────────────────────
...
    PayoutRequest:
      type: object
      required: [id, amount, bank_name, bank_account, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
          description: Payout amount in MNT.
        bank_name:
          type: string
          example: Хаан банк
        bank_account:
          type: string
          example: "5012345678"
        status:
          type: string
          enum: [PENDING, PROCESSED, REJECTED]
        created_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true

    # ─── Reviews ──────────────────────────────────────────────
...
    Dispute:
      type: object
      required: [id, booking_id, raised_by, reason, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        booking_id:
          type: string
          format: uuid
        booking:
          $ref: '#/components/schemas/Booking'
        raised_by:
          type: string
          format: uuid
        reason:
          type: string
        status:
          type: string
          enum: [OPEN, RESOLVED]
        resolution:
          type: string
          enum: [REFUND_CUSTOMER, RELEASE_TASKER]
          nullable: true
        resolution_amount:
          type: integer
          nullable: true
        resolution_notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true

    # ─── Conversations / Messages ─────────────────────────────
...
    Message:
      type: object
      required: [id, conversation_id, sender_id, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get booking details
      description: Returns full booking details. Only accessible by the booking's customer or tasker.
      operationId: getBooking
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Booking details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /disputes/{id}:
    get:
      tags: [Disputes]
      summary: Get dispute details
      description: Returns dispute details. Accessible by the dispute raiser, the other booking party, and admins.
      operationId: getDispute
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Dispute details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Conversations ──────────────────────────────────────────
...
  /admin/disputes:
    get:
      tags: [Admin]
      summary: List open disputes
      description: Returns paginated list of open disputes for admin resolution.
      operationId: adminListDisputes
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of open disputes.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dispute'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/disputes/{id}:
    get:
      tags: [Admin]
      summary: Get dispute details with evidence context
      description: |
        Returns full dispute details, related booking context, and recent conversation messages
        between booking participants for moderation review.
      operationId: adminGetDispute
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Dispute detail payload for admin review.
          content:
            application/json:
              schema:
                type: object
                properties:
                  dispute:
                    $ref: '#/components/schemas/Dispute'
                  booking:
                    $ref: '#/components/schemas/Booking'
                  conversation_id:
                    type: string
                    format: uuid
                    nullable: true
                  evidence_messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/disputes/{id}/resolve:
    post:
      tags: [Admin]
      summary: Resolve a dispute
      description: |
        Admin resolves a dispute by selecting customer-favor, tasker-favor, or escalation outcome.
        Post-MVP monetization may extend this endpoint with refund/release amount handling.
        Requires Idempotency-Key header.
      operationId: adminResolveDispute
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                  enum: [RESOLVE_CUSTOMER, RESOLVE_TASKER, ESCALATE]
                notes:
                  type: string
                  maxLength: 2000
                  description: Admin notes on the resolution decision.
      responses:
        '200':
          description: Dispute resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Dispute is already resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/moderation/strike-policy:
    get:
      tags: [Admin]
      summary: Get strike and suspension policy
      description: Returns current moderation policy used for tasker strike suspension logic.
      operationId: adminGetStrikePolicy
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current strike policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrikePolicy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Admin]
      summary: Update strike and suspension policy
      description: |
        Updates moderation policy values used for suspension thresholds and repeat-offense escalation.
      operationId: adminUpdateStrikePolicy
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrikePolicyUpdateRequest'
      responses:
        '200':
          description: Strike policy updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrikePolicy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/payouts/pending:
    get:
      tags: [Admin]
      summary: List pending payouts
      deprecated: true
      description: Post-MVP deferred endpoint. Returns paginated pending payout requests.
      operationId: adminListPendingPayouts
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of pending payouts.
          content:
            application/json:
              schema:
                type: object
                required: [data, cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PayoutRequest'
                  cursor:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Monetization is deferred for the current MVP phase.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/payouts/{id}/process:
    post:
      tags: [Admin]
      summary: Mark payout as processed
      deprecated: true
      description: |
        Post-MVP deferred endpoint.
        Admin marks a payout request as processed after completing the bank transfer.
        This debits the tasker's wallet and creates a ledger entry.
        Requires Idempotency-Key header.
      operationId: adminProcessPayout
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '200':
          description: Payout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Payout is not in PENDING status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Monetization is deferred for the current MVP phase.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...